requirements.txt
oracledb==2.2.1
pyodbc==5.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31